Полный тест блока "Создание пользователя"

+++ src/test/java/TestCreateUser.java
@@ -1,28 +1,38 @@
-import model.DeleteUser;
-import user.CreateUserApi;
-import model.CreateUser;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import org.junit.AfterClass;
+import model.CreateUser;
+import model.DeleteUser;
+import org.junit.After;
 import org.junit.Test;
+import user.CreateUserApi;
 import user.DeleteUserApi;
 
-import static java.util.function.Predicate.not;
-import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.containsString;
+import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.core.IsEqual.equalTo;
 
 
 public class TestCreateUser {
-    private static String email = "andrianovpa1@gmail.com";
-    private static String password = "123456781";
-    private static String name = "Pavel1";
+    private static String email = "andrianovpa@gmail.com";
+    private static String password = "12345678";
+    private static String name = "Pavel";
     private static String accessToken;
 
-    @AfterClass
-    public static void deleteUser() {
-        DeleteUserApi deleteUserApi = new DeleteUserApi();
-        DeleteUser deleteUser = new DeleteUser();
-        deleteUserApi.deleteUser(accessToken);
+    @After
+    public void deleteUser() {
+        try {
+            // Проверяем, не равен ли accessToken null
+            if (accessToken != null) {
+                DeleteUserApi deleteUserApi = new DeleteUserApi();
+                deleteUserApi.deleteUser(accessToken);
+                System.out.println("Пользователь успешно удалён");
+            } else {
+                System.out.println("Удаление пользователя не требуется");
+            }
+        } catch (Exception e) {
+            // Обработка исключений
+            System.err.println("Ошибка при удалении пользователя: " + e.getMessage());
+        }
     }
 
 
@@ -42,34 +52,32 @@ public class TestCreateUser {
                 .body("user.name", equalTo(name))
                 .extract().path("accessToken");
     }
+
+    @Test
+    @DisplayName("Проверка создания пользователя ранее созданного в системе") // имя теста
+    @Description("Направялется запрос на создание пользователя с валидными данными два раза и после второго раза проверяется код и тело ответа") // описание теста
+    public void negativeCreateUserTest() {
+
+        CreateUserApi createUserApi = new CreateUserApi();
+        CreateUser createUser = new CreateUser(email, password, name);
+        accessToken = createUserApi.createUser(createUser).then().extract().path("accessToken");
+        createUserApi.createUser(createUser)
+                .then().statusCode(403).assertThat()
+                .body("success", equalTo(false))
+                .body("message", equalTo("User already exists"));
+    }
+
+@Test
+@DisplayName("Проверка создания пользователя ,без обязательного поля") // имя теста
+@Description("Направялется запрос на создание пользователя без тега name") // описание теста
+public void negativeCreateUserWithoutFieldTest() {
+
+    CreateUserApi createUserApi = new CreateUserApi();
+    CreateUser createUser = new CreateUser(email, password, null);
+    accessToken = createUserApi.createUser(createUser)
+            .then().statusCode(403).assertThat()
+            .body("success", equalTo(false))
+            .body("message", equalTo("Email, password and name are required fields"))
+            .extract().path("accessToken");
 }
-//    @Test
-//    @DisplayName("Негативная проверка создания курьера 1") // имя теста
-//    @Description("Направляется запрос на создание курьера с уже зарегестрированными данными") // описание теста
-//    public void negativeCreateCourierTest_1() {
-//
-//        CreateCourierApi createCourierApi = new CreateCourierApi();
-//        CreateCourier courier = new CreateCourier(login, password, firstName);
-//        createCourierApi.createCourier(courier).then().statusCode(409).assertThat().body("message", equalTo("Этот логин уже используется"));
-//    }
-//
-//    @Test
-//    @DisplayName("Негативная проверка создания курьера 2") // имя теста
-//    @Description("Направляется запрос на создание курьера без поля login") // описание теста
-//    public void negativeCreateCourierTest_2() {
-//
-//        CreateCourierApi createCourierApi = new CreateCourierApi();
-//        CreateCourier courier = new CreateCourier(null, password, firstName);
-//        createCourierApi.createCourier(courier).then().statusCode(400).assertThat().body("message", equalTo("Недостаточно данных для создания учетной записи"));
-//    }
-//
-//    @Test
-//    @DisplayName("Негативная проверка создания курьера 3") // имя теста
-//    @Description("Направляется запрос на создание курьера без поля password") // описание теста
-//    public void negativeCreateCourierTest_3() {
-//
-//        CreateCourierApi createCourierApi = new CreateCourierApi();
-//        CreateCourier courier = new CreateCourier(login, null, firstName);
-//        createCourierApi.createCourier(courier).then().statusCode(400).assertThat().body("message", equalTo("Недостаточно данных для создания учетной записи"));
-//    }
-//}
+}
